+++ changelog.md
@@ -1,12 +1,14 @@
 ### Version 0.1.1
-- Dialog with instructions updated
+- Possibility of adding custom Animation
+- Adjustable width of Animated Banner's parent
 ### Version 0.1.0
 - Custom animation params feature added
+- Dialog with instructions updated
 ### Version 0.0.4
 - Dialog with instructions added
 ### Version 0.0.3
 - Animated Banner params such custom color and custom background, width and height, can be edited
-- color picker component added
+- Color picker component added
 ### Version 0.0.2
 - Animated Banner web component sample params preview
 ### Version 0.0.1